+++ src/ui/settingsMapper.ts
@@ -29,6 +29,9 @@ export const mapSettingsToYarleOptions = (): YarleOptions => {
         keepOriginalHtml: store.get('keepOriginalHtml') as boolean,
         currentTemplate: store.get('currentTemplate') as string,
         resourcesDir: store.get('resourcesDir') as string,
+        trimStartingTabs: store.get('trimStartingTabs') as boolean,
+        convertPlainHtmlNewlines: store.get('convertPlainHtmlNewlines') as boolean,
+        encryptionPasswords: ((store.get('encryptionPasswords') as string)||'').split(',').map(pwd => pwd.trim()),
         nestedTags: {
             separatorInEN: store.get('nestedTags.separatorInEN') as string,
             replaceSeparatorWith: store.get('nestedTags.replaceSeparatorWith') as string,