+++ src/utils/templates/remove-functions/remove-evernotelink-placeholder.ts
@@ -1,0 +1,7 @@
+import * as T from '../placeholders/evernotelink-placeholders';
+
+import { removePlaceholder } from './remove-placeholder';
+
+export const removeEvernoteLinkPlaceholder = (text: string): string => {
+  return removePlaceholder(text, T);
+};