+++ src/utils/turndown-service.ts
@@ -16,6 +16,7 @@ import { taskListRule } from './turndown-rules/task-list-rule';
 import { removeNewlines } from './remove-newlines';
 import { tanaTableBlock, tanaTableColBlock, tanaTableRowBlock } from './../constants';
 import { underlineRule } from './turndown-rules/underline-rule';
+import { encryptRule } from './turndown-rules/encrypt-rule';
 
 export const getTurndownService = (yarleOptions: YarleOptions) => {
     /* istanbul ignore next */
@@ -41,6 +42,7 @@ export const getTurndownService = (yarleOptions: YarleOptions) => {
     turndownService.addRule('list', taskListRule);
     turndownService.addRule('italic', italicRule);
     turndownService.addRule('underline', underlineRule);
+    turndownService.addRule('encrypt', encryptRule);
     if (yarleOptions.outputFormat === OutputFormat.LogSeqMD) {
         turndownService.addRule('logseq_hr', {
                 filter: ['hr'],